Separated package into multiple files for ease of modification. Updated README.md. Started work on UpgradeManager to handle interleaving package and app upgrades.
