+++ src/DtronixPackage.Tests/IntegrationTests/PackageUpgradeCallback.cs
@@ -11,17 +11,19 @@ namespace DtronixPackage.Tests.IntegrationTests
         private readonly Func<DtronixPackageCallbackUpgradeCallbackEventArgs, Task<bool>> Upgrading;
 
         public PackageUpgradeCallback(
-            Version version, 
+            Version packageDependentVersion,
+            Version appVersion, 
             Func<DtronixPackageCallbackUpgradeCallbackEventArgs, Task<bool>> upgrading)
-            : base(version)
+            : base(packageDependentVersion, appVersion)
         {
             Upgrading = upgrading;
         }
 
         public PackageUpgradeCallback(
-            Version version, 
+            Version packageDependentVersion,
+            Version appVersion,
             Func<DtronixPackageCallbackUpgradeCallbackEventArgs, Task> upgrading)
-            : base(version)
+            : base(packageDependentVersion, appVersion)
         {
             Upgrading = args =>
             {
