+++ src/DtronixPackage/Package.cs
@@ -9,7 +9,6 @@ using System.Linq;
 using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
-using System.Windows.Media.Imaging;
 using DtronixPackage.Logging;
 using DtronixPackage.RecursiveChangeNotifier;
 using DtronixPackage.Upgrades;
@@ -21,7 +20,7 @@ namespace DtronixPackage
     /// </summary>
     /// <typeparam name="TContent"></typeparam>
     [SuppressMessage("ReSharper", "StaticMemberInGenericType")]
-    public abstract class Package<TContent> : IPackage 
+    public abstract partial class Package<TContent> : IPackage 
         where TContent : PackageContent, new()
     {
         private readonly string _appName;
@@ -32,12 +31,11 @@ namespace DtronixPackage
         private FileStream _openPackageStream;
 
         private string _lockFilePath;
-        private readonly SemaphoreSlim _packageOperationSemaphore = new SemaphoreSlim(1, 1);
-        private readonly List<ChangelogEntry> _changelog = new List<ChangelogEntry>();
+        private readonly SemaphoreSlim _packageOperationSemaphore = new(1, 1);
+        private readonly List<ChangelogEntry> _changelog = new();
         private readonly Timer _autoSaveTimer;
         private bool _autoSaveEnabled;
-        private readonly Dictionary<object, ChangeListener> _monitorListeners 
-            = new Dictionary<object, ChangeListener>();
+        private readonly Dictionary<object, ChangeListener> _monitorListeners = new();
         private int _autoSavePeriod = 60 * 1000;
         private int _autoSaveDueTime = 60 * 1000;
         private bool _disposed;
@@ -108,7 +106,7 @@ namespace DtronixPackage
         /// Contains a list of upgrades which will be performed on older versions of packages.
         /// Add to this list to include additional upgrades.  Will execute in the order listed.
         /// </summary>
-        protected List<PackageUpgrade> Upgrades { get; } = new List<PackageUpgrade>();
+        protected List<PackageUpgrade> Upgrades { get; } = new();
 
         /// <summary>
         /// Internal calculated property used to get the current time.
@@ -146,6 +144,9 @@ namespace DtronixPackage
             }
         }
 
+        /// <summary>
+        /// Stored content of the package.
+        /// </summary>
         public TContent Content { get; }
 
         /// <summary>
@@ -210,277 +211,12 @@ namespace DtronixPackage
         }
 
         /// <summary>
-        /// Method called when opening a package file and to run the application specific opening code.
-        /// By default, will read content.json and copy the contents to the Content object.
+        /// Override with a method to specify a temp file location.
         /// </summary>
-        /// <param name="reader">Package reader.</param>
-        /// <returns>True of successful opening. False otherwise.</returns>
-        protected abstract Task<bool> OnRead(PackageReader reader);
-
-        /// <summary>
-        /// Method called when saving.  By default, will save the Content object to contents.json
-        /// </summary>
-        /// <param name="writer">Package writer.</param>
-        protected abstract Task OnWrite(PackageWriter writer);
-
+        /// <param name="fileName">Filename to include in the temp file path result.</param>
+        /// <returns>New temp file path location.</returns>
         protected abstract string OnTempFilePathRequest(string fileName);
 
-        /// <summary>
-        /// Opens a package.
-        /// </summary>
-        /// <param name="path">Package path to open.</param>
-        /// <param name="openReadOnly">
-        /// Specifies that the package will be open in read-only mode and can not be saved back
-        /// to the source file.
-        /// </param>
-        /// <param name="cancellationToken">Cancellation token for opening.</param>
-        /// <returns>True if opening was successful, otherwise false.</returns>
-        public async Task<PackageOpenResult> Open(
-            string path,
-            bool openReadOnly = false, 
-            CancellationToken cancellationToken = default)
-        {
-            if (_openPackageStream != null)
-                throw new InvalidOperationException("Can not open package when a package is already open");
-
-            await _packageOperationSemaphore.WaitAsync(cancellationToken);
-            var isMonitorEnabledOriginalValue = IsMonitorEnabled;
-            PackageOpenResult returnValue = null;
-
-            try
-            {
-                SavePath = path;
-                _lockFilePath = SavePath + ".lock";
-                IsReadOnly = openReadOnly;
-
-                // Read the lock file.
-                if (!openReadOnly)
-                {
-                    var (lockExists, lockFile) = await LockExists();
-
-                    if(lockExists)
-                        return returnValue = new PackageOpenResult(PackageOpenResultType.Locked, null, lockFile);
-                }
-
-                try
-                {
-                    // Copy the package into memory for use.
-                    // Retain the package lock to ensure the file does not get changed.
-                    _openPackageStream = new FileStream(
-                        path,
-                        FileMode.Open,
-                        openReadOnly ? FileAccess.Read : FileAccess.ReadWrite,
-                        FileShare.ReadWrite);
-                }
-                catch (FileNotFoundException e)
-                {
-                    return returnValue = new PackageOpenResult(PackageOpenResultType.FileNotFound, e);
-                }
-                catch (DirectoryNotFoundException e)
-                {
-                    return returnValue = new PackageOpenResult(PackageOpenResultType.FileNotFound, e);
-                }
-                catch (DriveNotFoundException e)
-                {
-                    return returnValue = new PackageOpenResult(PackageOpenResultType.FileNotFound, e);
-                }
-                catch (PathTooLongException e)
-                {
-                    return returnValue = new PackageOpenResult(PackageOpenResultType.FileNotFound, e);
-                }
-                catch (UnauthorizedAccessException e)
-                {
-                    return returnValue = new PackageOpenResult(PackageOpenResultType.PermissionFailure, e);
-                }
-                catch (IOException e) when ((e.HResult & 0x0000FFFF) == 32)
-                {
-                    return returnValue = new PackageOpenResult(PackageOpenResultType.Locked, e);
-                }
-                catch (Exception e)
-                {
-                    Logger?.Error(e, "Unknown error while opening file.");
-                    return returnValue = new PackageOpenResult(PackageOpenResultType.UnknownFailure, e);
-                }
-
-                // Read the entire contents to memory for usage in archive Update mode.
-                Stream openPackageStreamCopy = new MemoryStream();
-                await _openPackageStream.CopyToAsync(openPackageStreamCopy, cancellationToken);
-
-                if (openReadOnly)
-                {
-                    _openPackageStream.Close();
-                    _openPackageStream = null;
-                }
-
-                try
-                {
-                    _openArchive = new ZipArchive(openPackageStreamCopy, ZipArchiveMode.Update);
-
-                    // Read the package version number
-                    var packageVersionEntry = _openArchive.Entries.FirstOrDefault(f => f.FullName == "version");
-
-                    if (packageVersionEntry == null)
-                    {                   
-                        // If the package version is not set, this is an older file which will need to be upgraded appropriately.
-                        _openPackageVersion = new Version(0, 0, 0);
-                    }
-                    else
-                    {
-                        using var packageVersionReader = new StreamReader(packageVersionEntry.Open());
-                        _openPackageVersion = new Version(await packageVersionReader.ReadToEndAsync());
-                    }
-
-                    // Don't allow opening if the package file was saved with a more recent version of the packager.
-                    if (PackageVersion < _openPackageVersion)
-                        return returnValue = new PackageOpenResult(PackageOpenResultType.IncompatiblePackageVersion, PackageVersion);
-
-                    // Read the version information in the application directory.
-                    var versionEntity = _openArchive.Entries.FirstOrDefault(f => f.FullName == _appName + "/version");
-
-                    // If the versionEntity is empty, this usually means that this package was not created by the same
-                    // application currently opening the package.
-                    if (versionEntity == null)
-                    {
-
-                        // Check to see if there is a version file in the entire package.  If there is, there is a high
-                        // probability that this is a Package file, just used by another application's package system.
-                        if(_openArchive.Entries.Any(f => f.FullName.EndsWith("/version")))
-                            return returnValue = new PackageOpenResult(PackageOpenResultType.IncompatibleApplication);
-                        else
-                            return returnValue = new PackageOpenResult(PackageOpenResultType.Corrupted);
-                    }
-
-                    using (var appVersionReader = new StreamReader(versionEntity.Open()))
-                        Version = new Version(await appVersionReader.ReadToEndAsync());
-
-                    // Don't allow opening of newer packages on older applications.
-                    if (AppVersion < Version)
-                        return returnValue = new PackageOpenResult(PackageOpenResultType.IncompatibleVersion, Version);
-                }
-                catch(Exception e)
-                {
-                    return returnValue = new PackageOpenResult(PackageOpenResultType.Corrupted, e, Version);
-                }
-
-                // Perform any required package upgrades.
-                if (_openPackageVersion < PackageVersion)
-                {
-                    var packageUpgrades = new PackageUpgrade[]
-                    {
-                        new PackageUpgrade_1_1_0(),
-                    };
-
-                    var upgradeResult = await ApplyUpgrades(packageUpgrades, true, _openPackageVersion);
-
-                    // If the result is not null, the upgrade failed.
-                    if(upgradeResult != null)
-                        return returnValue = upgradeResult;
-                }
-
-                // Try to open the modified log file. It may not exist in older versions.
-                var changelogEntry =
-                    _openArchive.Entries.FirstOrDefault(f => f.FullName == _appName + "/changelog.json");
-
-                if (changelogEntry != null)
-                {
-                    try
-                    {
-                        _changelog.Clear();
-                        await using var stream = changelogEntry.Open();
-                        var logItems = await JsonSerializer.DeserializeAsync<ChangelogEntry[]>(
-                            stream, 
-                            null, 
-                            cancellationToken);
-
-                        foreach (var changelogItem in logItems)
-                            _changelog.Add(changelogItem);
-                        
-                    }
-                    catch (Exception e)
-                    {
-                        Logger?.Error(e, "Could not parse save log.");
-                        _changelog.Clear();
-                    }
-                }
-
-                // Perform upgrades at this time.
-                if (AppVersion > Version)
-                {
-                    // "Renames" existing contents of this application's directory to a backup directory 
-                    if (_preserveUpgrade)
-                    {
-                        var entries = _openArchive.Entries.ToArray();
-                        foreach (var entry in entries)
-                        {
-                            // If there is a version miss-match, save the original files in a modified
-                            // directory to retrieve in-case of corruption.
-                            if (!entry.FullName.StartsWith(_appName + "/")) 
-                                continue;
-
-                            // Renames all the sub-files if there was a version mis-match on open.
-                            var newName = _appName + $"-backup-{Version}/{entry.FullName}";
-
-                            var saveEntry = _openArchive.CreateEntry(newName);
-
-                            // Copy the last write times to be accurate.
-                            saveEntry.LastWriteTime = entry.LastWriteTime;
-
-                            // Copy the streams.
-                            await using var openedArchiveStream = entry.Open();
-                            await using var saveEntryStream = saveEntry.Open();
-                            await openedArchiveStream.CopyToAsync(saveEntryStream, cancellationToken);
-                        }
-                    }
-
-                    var upgradeResult = await ApplyUpgrades(Upgrades, false, Version);
-
-                    // If the result is not null, the upgrade failed.
-                    if(upgradeResult != null)
-                        return returnValue = upgradeResult;
-                }
-
-                // If we are not in read-only mode and we are using lock files, create a lock file
-                // and if it failed, stop opening.
-                if (!openReadOnly && _useLockFile && !await SetLockFile())
-                {
-                    return returnValue = new PackageOpenResult(PackageOpenResultType.Locked, Version);
-                }
-
-                OpenTime = DateTime.Now;
-                AutoSavePath = null;
-
-                var reader = new PackageReader(_openArchive, JsonSerializerOptions, _appName);
-
-                // Disable monitor notifications while reading.
-                IsMonitorEnabled = false;
-
-                // Perform the overridden reading operations.
-                var openResult = await OnRead(reader);
-
-                return returnValue = openResult
-                    ? PackageOpenResult.Success
-                    : new PackageOpenResult(PackageOpenResultType.ReadingFailure, Version);
-
-            }
-            catch (Exception e)
-            {
-                return returnValue = new PackageOpenResult(PackageOpenResultType.UnknownFailure, e, Version);
-            }
-            finally
-            {
-                // Perform cleanup if the open was not successful.
-                // Do not fire the close event since it was not successfully opened in the first place.
-                if (returnValue != PackageOpenResult.Success)
-                    CloseInternal(false);
-
-                // Restore the original monitor value.
-                IsMonitorEnabled = isMonitorEnabledOriginalValue;
-
-                _packageOperationSemaphore.Release();
-            }
-        }
-
-        
         /// <summary>
         /// Updates the currently open lock file with the current username & DateTime.
         /// </summary>
@@ -555,416 +291,6 @@ namespace DtronixPackage
                 // The lock file is locked by another process or the file was malformed.
                 return (true, null);
             }
-        }
-
-        /// <summary>
-        /// Saves the current package to the specified path.
-        /// </summary>
-        /// <param name="path">Path to output the package.</param>
-        public async Task<PackageSaveResult> Save(string path)
-        {
-            // Check the path if we are not auto-saving.
-            if (string.IsNullOrWhiteSpace(path))
-                throw new ArgumentNullException(nameof(path));
-
-            Logger?.ConditionalTrace("Requesting _packageOperationSemaphore");
-            // Ensure we are not already saving.
-            await _packageOperationSemaphore.WaitAsync();
-
-            Logger?.ConditionalTrace("Acquired _packageOperationSemaphore");
-
-
-
-            // Determine if we have changed save the save path location since the last save.
-            if (path != SavePath)
-            {
-                // If the package stream is open and we are changing the path, close the old stream.
-                if (_openPackageStream != null)
-                {
-                    _openPackageStream.Close();
-                    _openPackageStream = null;
-                }
-
-                SavePath = path;
-
-                // If the package is new and not saved yet, a lock file has not been created.
-                if (_lockFile != null)
-                {
-                    // If the _lockFile variable is set, and the path has changed, this means 
-                    // that there is an extra lock file which needs to be removed.  Try to remove it.
-                    try
-                    {
-                        _lockFile.Close();
-                        _lockFile = null;
-                        File.Delete(_lockFilePath);
-                    }
-                    catch (Exception e)
-                    {
-                        Logger?.Info(e, $"Unable to remove lock file at: {_lockFilePath}.");
-                    }
-                }
-
-                _lockFilePath = path + ".lock";
-            }
-
-            // Lock file creation.
-            try
-            {
-                if (_useLockFile && _lockFile == null)
-                {
-                    var (lockExists, lockFile) = await LockExists();
-
-                    if(lockExists)
-                        return new PackageSaveResult(PackageSaveResultType.Locked, null, lockFile);
-
-                    if (!await SetLockFile())
-                        return new PackageSaveResult(PackageSaveResultType.Locked);
-                }
-                
-                return await SaveInternal(false);
-            }
-            catch (Exception e)
-            {
-                return new PackageSaveResult(PackageSaveResultType.Failure, e);
-            }
-        }
-
-        /// <summary>
-        /// Core save functionality.
-        /// </summary>
-        /// <param name="autoSave">
-        /// Set to true to bypass saving to the SavePath and instead save the current package to the temp directory.
-        /// </param>
-        /// <returns>Result of saving</returns>
-        private async Task<PackageSaveResult> SaveInternal(bool autoSave)
-        {
-            Logger?.ConditionalTrace($"SaveInternal(autoSave:{autoSave})");
-            PackageSaveResult returnValue = null;
-            try
-            {
-                // TODO: Reuse existing archive when saving from a read-only package.  Prevents duplication of the MS.
-                var saveArchiveMemoryStream = new MemoryStream();
-
-                using (var archive = new ZipArchive(saveArchiveMemoryStream, ZipArchiveMode.Create, true))
-                {
-                    var writer = new PackageWriter(archive, JsonSerializerOptions, _appName);
-
-                    // Write application version file
-                    await using (var packageVersionStream = writer.CreateEntityStream("version", false))
-                    {
-                        await using var packageVersionStreamWriter = new StreamWriter(packageVersionStream);
-                        await packageVersionStreamWriter.WriteAsync(PackageVersion.ToString());
-                    }
-
-                    await OnWrite(writer);
-
-                    // Write package version file
-                    await writer.Write("version", AppVersion.ToString());
-
-                    var log = new ChangelogEntry
-                        (autoSave ? ChangelogEntryType.AutoSave : ChangelogEntryType.Save,
-                        Username,
-                        ComputerName,
-                        CurrentDateTimeOffset);
-
-                    // Update the save log.
-                    _changelog.Add(log);
-
-                    // Write the save log.
-                    await writer.WriteJson("changelog.json", _changelog);
-
-                    // If this is an auto save, we do not want to continually add auto save logs.
-                    if (autoSave)
-                        _changelog.Remove(log);
-
-                    if (_openArchive != null)
-                    {
-                        foreach (var openedArchiveEntry in _openArchive.Entries)
-                        {
-                            if (writer.FileList.Contains(openedArchiveEntry.FullName))
-                                continue;
-
-                            var saveEntry = archive.CreateEntry(openedArchiveEntry.FullName);
-
-                            // Copy the last write times to be accurate.
-                            saveEntry.LastWriteTime = openedArchiveEntry.LastWriteTime;
-
-                            // Copy the streams.
-                            await using var openedArchiveStream = openedArchiveEntry.Open();
-                            await using var saveEntryStream = saveEntry.Open();
-                            await openedArchiveStream.CopyToAsync(saveEntryStream);
-                        }
-                    }
-                    
-                }
-
-                // Only save a backup package if we are not performing an auto save.
-                if (SaveBackupPackage && !autoSave)
-                {
-                    Logger?.ConditionalTrace("Saving backup package.");
-                    var bakPackage = SavePath + ".bak";
-                    try
-                    {
-                        Logger?.ConditionalTrace($"Checking for existence of {bakPackage} backup .");
-                        if (File.Exists(bakPackage))
-                        {
-                            Logger?.ConditionalTrace("Found backup package.");
-                            File.SetAttributes(bakPackage, FileAttributes.Normal);
-                            File.Delete(bakPackage);
-                            Logger?.ConditionalTrace("Deleted backup package.");
-                        }
-
-                        // Close the lock held on the file.  On initial save, there is nothing to close.
-                        _openPackageStream?.Close();
-                        _openPackageStream = null;
-
-                        Logger?.ConditionalTrace("Closed _openPackageStream.");
-                        Logger?.ConditionalTrace($"Checking for existence of existing save package {SavePath}");
-
-                        if (File.Exists(SavePath))
-                        {
-                            File.Move(SavePath, bakPackage);
-                            Logger?.ConditionalTrace($"Found existing save package and renamed to {bakPackage}");
-                        }
-                    }
-                    catch (Exception e)
-                    {
-                        return returnValue = new PackageSaveResult(PackageSaveResultType.Failure, e);
-                    }
-                }
-
-                if (!autoSave)
-                {
-                    // Create a new package/overwrite existing if a backup was not created.
-                    // Retain the lock on the package.
-                    if (_openPackageStream == null)
-                    {
-                        // Create a new source package.
-                        _openPackageStream = new FileStream(SavePath, FileMode.Create, FileAccess.ReadWrite, FileShare.Read);
-                    }
-                    else
-                    {
-                        // Set the package file length to 0 to reset for writing.
-                        _openPackageStream.Position = 0;
-                        _openPackageStream.SetLength(0);
-                    }
-                }
-
-                if (autoSave)
-                {
-                    Logger?.ConditionalTrace($"Creating AutoSave package {AutoSavePath}.");
-                }
-
-                var destinationStream = autoSave ? File.Create(AutoSavePath) : _openPackageStream;
-
-                if (autoSave)
-                {
-                    Logger?.ConditionalTrace($"Created AutoSave package {AutoSavePath}.");
-                }
-
-                if (destinationStream == null)
-                {
-                    Logger?.ConditionalTrace("Could not create/use destination stream.");
-                    return returnValue = new PackageSaveResult(PackageSaveResultType.Failure);
-                }
-
-                saveArchiveMemoryStream.Seek(0, SeekOrigin.Begin);
-
-                // Copy and flush the contents of the save file.
-                await saveArchiveMemoryStream.CopyToAsync(destinationStream);
-                await destinationStream.FlushAsync();
-
-                // If we were auto-saving, close the destination save and return true.
-                // We do not want to effect the current state of the rest of the package.
-                if (autoSave)
-                {
-                    destinationStream.Close();
-                    Logger?.ConditionalTrace("Closed AutoSave destination stream.");
-                    return returnValue = PackageSaveResult.Success;
-                }
-
-                saveArchiveMemoryStream.Seek(0, SeekOrigin.Begin);
-
-                // This implicitly closes the inner stream.
-                _openArchive?.Dispose();
-
-                // Save the new zip archive stream to the opened archive stream.
-                _openArchive = new ZipArchive(saveArchiveMemoryStream, ZipArchiveMode.Read, true);
-
-                Version = AppVersion;
-
-                return returnValue = PackageSaveResult.Success;
-            }
-            catch (Exception e)
-            {
-                // Catch all for anything going completely crazy.
-                return returnValue = new PackageSaveResult(PackageSaveResultType.Failure, e);
-            }
-            finally
-            {
-                // If we were successful in saving, set the modified variable to false.
-                // If we are auto-saving, don't change these variables since the package still needs
-                // to actually be saved to the destination.
-                if (returnValue == PackageSaveResult.Success && autoSave == false)
-                {
-                    IsContentModified = false;
-
-                    // Since the package was saved, the package is no longer in read-only mode.
-                    IsReadOnly = false;
-                }
-
-                Logger?.ConditionalTrace($"InternalSave return: {returnValue}");
-                Logger?.ConditionalTrace("Released _packageOperationSemaphore");
-                _packageOperationSemaphore.Release();
-            }
-        }
-
-        /// <summary>
-        /// Sets the due time and period for auto-saves.  See <see cref="System.Threading.Timer"/>. If either the
-        /// time or the period are greater than -1, the auto save is enabled otherwise, it is disabled.
-        /// </summary>
-        /// <param name="dueTime">
-        /// The amount of time to delay before the invoking auto save, in milliseconds.
-        /// Specify Infinite to prevent auto save from restarting. Specify zero (0) to trigger an auto save immediately.
-        /// </param>
-        /// <param name="period">
-        /// The time interval between auto save invocations, in milliseconds.
-        /// Specify Infinite to disable periodic auto saves.
-        /// </param>
-        /// <param name="enable">Set to true to enable saving now if configured to run.</param>
-        public async Task ConfigureAutoSave(int dueTime, int period, bool enable)
-        {
-            Logger?.ConditionalTrace($"ConfigureAutoSave(dueTime:{dueTime}, period:{period})");
-
-            if (dueTime < -1)
-                throw new ArgumentOutOfRangeException(nameof(dueTime), "Value must be -1 or higher.");
-
-            if (period < -1)
-                throw new ArgumentOutOfRangeException(nameof(period), "Value must be -1 or higher.");
-
-            _autoSaveDueTime = dueTime;
-            _autoSavePeriod = period;
-
-            if (dueTime == 0)
-            {
-                Logger?.ConditionalTrace("Running auto save immediately.");
-                // If we call for it to save once immediately, call now.
-                await AutoSave(true);
-
-                // If this is a periodical save, push the due time up to the period time since we 
-                // just synchronously called the AutoSaveElapsed method.
-                if (period > -1)
-                {
-                    Logger?.ConditionalTrace($"Forward setting since AutoSave was just called. _autoSaveDueTime = {period}");
-                    _autoSaveDueTime = period;
-                }
-                else
-                {
-                    // If we are not recurring, then do nothing else.
-                    return;
-                }
-            }
-
-            // If the auto save is configured to run at least once, enable the auto save.  Otherwise disable.
-            AutoSaveEnabled = enable && (dueTime >= 0 || period >= 0);
-        }
-
-        private async Task AutoSave(bool synchronous)
-        {
-            Logger?.ConditionalTrace($"AutoSave(synchronous:{synchronous})");
-
-            // If the AutoSave path has not been set, set it now.
-            AutoSavePath ??= OnTempFilePathRequest(OpenTime.ToString("yyyy-MM-dd-T-HH-mm-ss-fff") + ".sav");
-
-            // Ensure that auto-save is still enabled & that the content has been modified.
-            if ((AutoSaveEnabled || synchronous) && IsDataModifiedSinceAutoSave)
-            {
-
-                Logger?.ConditionalTrace("Requesting _packageOperationSemaphore");
-                // Ensure we are not already saving.
-                await _packageOperationSemaphore.WaitAsync();
-                Logger?.ConditionalTrace("Acquired _packageOperationSemaphore");
-
-                var saveResult = await SaveInternal(true);
-
-                if (saveResult == PackageSaveResult.Success)
-                {
-                    // Set the internal variable to ensure that we do not auto-save more than what is required.
-                    IsDataModifiedSinceAutoSave = false;
-                }
-                else
-                {
-                    Logger?.Error(saveResult.Exception, $"Could not successfully auto-save package {saveResult.SaveResult}");
-                }
-            }
-        }
-
-        private void AutoSaveElapsed(object state)
-        {
-            Logger?.ConditionalTrace("AutoSaveElapsed()");
-            _ = AutoSave(false);
-        }
-
-        private void SetAutoSaveTimer()
-        {
-            // Update the current timer if it is running.
-            if (_autoSaveEnabled)
-            {
-                Logger?.ConditionalTrace($"_autoSaveDueTime = {_autoSaveDueTime}; _autoSavePeriod = {_autoSavePeriod}");
-                _autoSaveTimer.Change(_autoSaveDueTime, _autoSavePeriod);
-            }
-            else
-            {
-                Logger?.ConditionalTrace("_autoSaveDueTime = -1; _autoSavePeriod = -1");
-                _autoSaveTimer.Change(-1, -1);
-            }
-        }
-
-        private async Task<PackageOpenResult> ApplyUpgrades(
-            IList<PackageUpgrade> upgrades, 
-            bool packageUpgrade, 
-            Version compareVersion)
-        {
-            var currentVersion = packageUpgrade ? _openPackageVersion : Version;
-
-            foreach (var upgrade in upgrades.Where(upgrade => upgrade.Version > compareVersion))
-            {
-                try
-                {
-                    // Attempt to perform the upgrade
-                    if (!await upgrade.Upgrade(_openArchive))
-                    {
-                        // Upgrade soft failed, log it and notify the opener.
-                        Logger?.Error($"Unable to perform{(packageUpgrade ? " package" : " application")} upgrade of package to version {upgrade.Version}.");
-                        return new PackageOpenResult(PackageOpenResultType.UpgradeFailure, Version);
-                    }
-
-                    _changelog.Add(new ChangelogEntry(packageUpgrade
-                        ? ChangelogEntryType.PackageUpgrade
-                        : ChangelogEntryType.ApplicationUpgrade,
-                        Username, 
-                        ComputerName,
-                        CurrentDateTimeOffset)
-                    {
-                        Note = packageUpgrade 
-                            ? $"Package upgrade from {currentVersion} to {upgrade.Version}"
-                            : $"Application upgrade from {currentVersion} to {upgrade.Version}"
-                    });
-
-                    currentVersion = upgrade.Version;
-                }
-                catch (Exception e)
-                {
-                    // Upgrade hard failed.
-                    Logger?.Error(e, $"Unable to perform{(packageUpgrade ? " package" : " application")} upgrade of package to version {upgrade.Version}.");
-                    return new PackageOpenResult(PackageOpenResultType.UpgradeFailure, e, Version);
-                }
-
-                // Since we did perform an upgrade, set set that the package has been changed.
-                IsContentModified = true;
-            }
-
-            return null;
         }
 
         protected void MonitorRegister<T>(T obj)
